Completed

+++ app/Http/Controllers/ListingController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use Auth;
+use App\Models\User;
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
@@ -57,6 +59,10 @@ class ListingController extends Controller
 
    //Update Listing
    public function update(Request $request, Listing $listing){
+       //Make sure logged in user is owner
+       if($listing->user_id != auth()->id() ){
+           abort(403, 'Unathorized Action!');
+       }
     //    dd($request->file('logo'));
         $formFields = $request->validate([
             'company' => 'required',
@@ -80,9 +86,18 @@ class ListingController extends Controller
    }
 
    // Delete Listing
-   public function destroy(Listing $listing){        
+   public function destroy(Listing $listing){  
+       // Make sure logged in user is owner!
+       if($listing->user_id != auth()->id()){
+           abort(403, 'Unauthorized Action!');
+       }      
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing is deleted successfully!');
+   }
+
+   public function manage(){
+
+    return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
    }
 }
