+++ routes/web.php
@@ -42,24 +42,30 @@ use App\Http\Controllers\ListingController;
 
 Route::get('/', [ListingController::class, 'index']);
 // Show Create form
-Route::get('listings/create-job', [ListingController::class, 'create']);
+Route::get('listings/create-job', [ListingController::class, 'create'])->middleware('auth');
 //Store data listing
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 // Show Edit Form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 // Update Listing
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 //Delete Listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
+//Manage Listing
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 //Single Listing
-Route::get('/listings/{listing}', [ListingController::class, 'show']); 
+Route::get('/listings/{listing}', [ListingController::class, 'show'])->middleware('auth'); 
 
 // Show user register form
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
-Route::post('/users', [UserController::class, 'store']);
+Route::post('/users', [UserController::class, 'store'])->middleware('auth');
 // Log User Out
-Route::Post('/logout', [UserController::class, 'logout']);
+Route::Post('/logout', [UserController::class, 'logout'])->middleware('auth');
+//Show Login Form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+//Loggin User
+Route::Post('/users/authenticate', [UserController::class, 'authenticate']);
